refactor: rename module aliases in order to keep from name collision
- npm module can have `@ORG/NAME_MODULE` names. Module aliases are thus named
with `@@-` in order to prevent collision.
- `@nodekit/express-middleware-attach` is upgraded.

Fixes https://github.com/marmoym/marmoym/issues/131
