+++ package.json
@@ -47,7 +47,7 @@
     "typescript": "^3.1.6"
   },
   "dependencies": {
-    "@nodekit/express-middleware-attach": "git+https://github.com/nodekit-org/express-middleware-attach.git",
+    "@nodekit/express-middleware-attach": "git+https://github.com/nodekit-org/express-middleware-attach.git#v0.0.2",
     "@nodekit/express-route-mapper": "git+https://github.com/nodekit-org/express-route-mapper.git#v0.0.4",
     "@nodekit/node-logger": "git+https://github.com/nodekit-org/node-logger.git#v0.0.3",
     "bcrypt": "^1.0.3",
