+++ src/middlewares/app.middlewares.ts
@@ -2,14 +2,14 @@ import * as bodyParser from "body-parser";
 import cookieParser from 'cookie-parser';
 import { MiddlewareDefinition } from '@nodekit/express-middleware-attach';
 
-import { API } from '@models/ApiURL';
-import corsHandler from '@middlewares/corsHandler';
-import errorHandler from '@middlewares/errorHandler';
-import httpLogger from "@middlewares/httpLogger";
-import launchStatusChecker from '@middlewares/launchStatusChecker';
-import routeNoMatchHandler from '@middlewares/routeNoMatchHandler';
-import routers from '@routes/routers';
-import state from '@models/state';
+import { API } from '@@models/ApiURL';
+import corsHandler from '@@middlewares/corsHandler';
+import errorHandler from '@@middlewares/errorHandler';
+import httpLogger from '@@middlewares/httpLogger';
+import launchStatusChecker from '@@middlewares/launchStatusChecker';
+import routeNoMatchHandler from '@@middlewares/routeNoMatchHandler';
+import routers from '@@routes/routers';
+import state from '@@models/state';
 
 const middlewareDefinitions: MiddlewareDefinition[] = [
   {
