+++ src/routes/default/routes.default.ts
@@ -1,8 +1,8 @@
 import { Router, Request, Response } from 'express';
 
-import ApiURL from '@models/ApiURL';
-import * as DebugService from '@services/debugService';
-import HttpMethod from '@constants/HttpMethod';
+import ApiURL from '@@models/ApiURL';
+import * as DebugService from '@@services/debugService';
+import HttpMethod from '@@constants/HttpMethod';
 // import { Route } from '../routes';
 
 const routeMap = [
